Mention roughly how to disable the mouse showing the Unity launcher
This gets really annoying when trying to move the mouse cursor from the
right monitor to the left monitor. The default settings makes it 'stick'
at the left edge of the right monitor, unless you move it very quickly.

I never want to show the launcher with the mouse, so setting the Edge
Stop Velocity to the lowest possible value, and Launcher Reveal Edge
Responsiveness to the lowest possible value, improves things a lot.